+++ demo/app/demo.module.ts
@@ -24,8 +24,6 @@ import { EntrySelectDemoComponent } from './entry-select/entry-select-demo.compo
 import { AssetListDemoComponent } from './asset-list/asset-list.demo.component';
 import { AssetDemoComponent } from './asset/asset.demo.component';
 import { ListTransformsDemoComponent } from './list-transforms/list-transforms-demo.component';
-import { DatamanagerListDemoComponent } from './datamanager-list/datamanager-list.demo.component';
-import { ModelListDemoComponent } from './model-list/model-list.demo.component';
 import { DatetimeDemoComponent } from './form/datetime-demo.component';
 import { DemoDashboardComponent } from './dashboard.component';
 import { AssetSelectDemoComponent } from './asset-select/asset-select-demo.component';
@@ -83,9 +81,7 @@ demoRoutes.unshift(
     ActionbarDemoComponent,
     InteractionDemoComponent,
     AssetDemoComponent,
-    DatamanagerListDemoComponent,
     DatetimeDemoComponent,
-    ModelListDemoComponent,
     NotificationsDemoComponent,
     LoaderDemoComponent,
     MediumEditorDemoComponent,